Add files via upload
Problem Statement:
A bike-sharing system is a service in which bikes are made available for shared use to individuals on a short term basis for a price or free. Many bike share systems allow people to borrow a bike from a "dock" which is usually computer-controlled wherein the user enters the payment information, and the system unlocks it. This bike can then be returned to another dock belonging to the same system.
Goals and Objectives:
To find which variables are significant in predicting the demand for shared bikes.
How well those variables describe the bike demands.